@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { FormGroup, FormControl, Validators } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormControl, Validators } from '@angular/forms';
 
 import { AuthService } from 'src/app/services/auth.service';
 import { RegisterValidators } from '../validators/register-validators';
@@ -15,23 +15,23 @@ export class RegisterComponent {
 
   inSubmission = false;
 
-  name = new FormControl('', [Validators.required, Validators.minLength(3)]);
-  email = new FormControl(
+  name = new UntypedFormControl('', [Validators.required, Validators.minLength(3)]);
+  email = new UntypedFormControl(
     '',
     [Validators.required, Validators.email],
     [this.emailTaken.validate]
   );
-  age = new FormControl('', [
+  age = new UntypedFormControl('', [
     Validators.required,
     Validators.min(18),
     Validators.max(120),
   ]);
-  password = new FormControl('', [
+  password = new UntypedFormControl('', [
     Validators.required,
     Validators.pattern(/^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[a-zA-Z]).{8,}$/gm),
   ]);
-  confirm_password = new FormControl('', [Validators.required]);
-  phoneNumber = new FormControl('', [
+  confirm_password = new UntypedFormControl('', [Validators.required]);
+  phoneNumber = new UntypedFormControl('', [
     Validators.required,
     Validators.minLength(13),
     Validators.maxLength(13),
@@ -41,7 +41,7 @@ export class RegisterComponent {
   alertMsg = 'Please wait. Your account is being created.';
   alertColor = 'blue';
 
-  registerForm = new FormGroup(
+  registerForm = new UntypedFormGroup(
     {
       name: this.name,
       email: this.email,
