@@ -1,5 +1,10 @@
 import { Component, OnInit, ɵɵsetComponentScope } from '@angular/core';
-import { FormControl, FormGroup, Validators } from '@angular/forms';
+import {
+  FormControl,
+  UntypedFormControl,
+  UntypedFormGroup,
+  Validators,
+} from '@angular/forms';
 
 @Component({
   selector: 'app-upload',
@@ -12,8 +17,11 @@ export class UploadComponent implements OnInit {
   // hide the form until file is uploaded
   nextStep = false;
 
-  title = new FormControl('', [Validators.required, Validators.minLength(3)]);
-  uploadForm = new FormGroup({
+  title = new FormControl('', {
+    validators: [Validators.required, Validators.minLength(3)],
+    nonNullable: true,
+  });
+  uploadForm = new UntypedFormGroup({
     title: this.title,
   });
 
